8.bigquery-vertexai-mcp-connection/streamlit-app/requirements.txt
streamlit
google-cloud-aiplatform
google-auth
requests
python-dotenv